+++ web/assets/js/config.example.js
@@ -1,10 +1,11 @@
 // enter your personal api keys here, get your own (free) keys by using the links below
 
 window.BOSM = {
-	osmchaTok: '',					// https://osmcha.mapbox.com/user, current token
+	osmchaTok: '',					// https://osmcha.mapbox.com/user
 	owmKey: '',						// https://openweathermap.org
 	mapboxKey: 'pk.',				// https://www.mapbox.com/pricing
 	thuforKey: '',					// https://thunderforest.com/pricing
+	googleKey: '',					// https://cloud.google.com/maps-platform
 	trvllneApi: { u: '', p: '' },	// http://www.travelinedata.org.uk
-	bookingCom: ''			// https://spadmin.booking.com, affiliate links
+	bookingCom: ''					// https://spadmin.booking.com
 };